add identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,12 +51,12 @@ public class AuctionsController: ControllerBase
         return _mapper.Map<AuctionDto>(auctions);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: add identity
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
         _context.Auctions.Add(auction);
 
         var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -67,6 +68,7 @@ public class AuctionsController: ControllerBase
             new {auction.Id}, newAuction );
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult<AuctionDto>> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
     {
@@ -76,7 +78,8 @@ public class AuctionsController: ControllerBase
         
         if (auction == null) return NotFound();
 
-        // TODO: check Identity
+        if (auction.Seller != User.Identity.Name) return Forbid();
+
         auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDTO.Color ?? auction.Item.Color;
@@ -93,6 +96,7 @@ public class AuctionsController: ControllerBase
         return BadRequest("Problem save update change");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -100,7 +104,7 @@ public class AuctionsController: ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TODO: check identity
+        if (auction.Seller != User.Identity.Name) return Forbid();
         _context.Auctions.Remove(auction);
 
         // publish to bus
